refactor: UserDetailsImpl, ServiceImpl, JwtUtil 리팩토링
