+++ src/main/java/com/team11/shareoffice/global/security/UserDetailsImpl.java
@@ -1,28 +1,23 @@
-//package com.team11.shareoffice.global.security;
-//
-//
-//import com.team11.shareoffice.member.entity.Member;
-//import org.springframework.security.core.GrantedAuthority;
-//import org.springframework.security.core.authority.SimpleGrantedAuthority;
-//import org.springframework.security.core.userdetails.UserDetails;
-//
-//import java.util.ArrayList;
-//import java.util.Collection;
-//
-//public class UserDetailsImpl implements UserDetails {
-//    private final Member member;
-//    private final String username;
-//
-//    public UserDetailsImpl(Member member, String username) {
-//        this.member = member;
-//        this.username = username;
-//    }
-//
-//    public Member getMember() {
-//        return member;
-//    }
-//
-//    @Override
+package com.team11.shareoffice.global.security;
+
+
+import com.team11.shareoffice.member.entity.Member;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+
+public class UserDetailsImpl implements UserDetails {
+    private final Member member;
+    private final String nickname;
+
+    public UserDetailsImpl(Member member, String nickname) {
+        this.member = member;
+        this.nickname = nickname;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {return null;}
 //    public Collection<? extends GrantedAuthority> getAuthorities() {
 //        String authority = "ROLE_USER";
 //
@@ -32,34 +27,37 @@
 //
 //        return authorities;
 //    }
-//
-//    @Override
-//    public String getUsername() {
-//        return this.username;
-//    }
-//
-//    @Override
-//    public String getPassword() {
-//        return null;
-//    }
-//
-//    @Override
-//    public boolean isAccountNonExpired() {
-//        return false;
-//    }
-//
-//    @Override
-//    public boolean isAccountNonLocked() {
-//        return false;
-//    }
-//
-//    @Override
-//    public boolean isCredentialsNonExpired() {
-//        return false;
-//    }
-//
-//    @Override
-//    public boolean isEnabled() {
-//        return false;
-//    }
-//}
+    public Member getMember() {
+    return member;
+}
+
+    @Override
+    public String getUsername() {
+        return this.member.getNickname();
+    }
+
+    @Override
+    public String getPassword() {
+        return null;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return false;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return false;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return false;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return false;
+    }
+}
