+++ src/main/java/com/team11/shareoffice/global/security/UserDetailsServiceImpl.java
@@ -1,26 +1,34 @@
-//package com.team11.shareoffice.global.security;
-//
-//import com.team11.shareoffice.member.entity.Member;
-//import com.team11.shareoffice.member.repository.MemberRepository;
-//import lombok.RequiredArgsConstructor;
-//import org.springframework.security.core.userdetails.UserDetails;
-//import org.springframework.security.core.userdetails.UserDetailsService;
-//import org.springframework.security.core.userdetails.UsernameNotFoundException;
-//import org.springframework.stereotype.Service;
-//
-//@Service
-//@RequiredArgsConstructor
-//public class UserDetailsServiceImpl implements UserDetailsService {
-//
-//    private final MemberRepository memberRepository;
-//
+package com.team11.shareoffice.global.security;
+
+import com.team11.shareoffice.member.entity.Member;
+import com.team11.shareoffice.member.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserDetailsServiceImpl implements UserDetailsService {
+
+    private final MemberRepository memberRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        Member member = memberRepository.findByEmail(email).orElseThrow(
+                () -> new UsernameNotFoundException("사용자를 찾을 수 없습니다.")
+        );
+        return new UserDetailsImpl(member, member.getNickname());
+    }
+
 //    @Override
 //    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
 //        //이 부분 코드 해결하고 주석 풀면 됩니다!
-////        Member member = memberRepository.findByEmail(username)
-////                .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
-////        return new UserDetailsImpl(member, member.getUsername());
+//        Member member = memberRepository.findByEmail(username)
+//                .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
+//        return new UserDetailsImpl(member, member.getUsername());
 //        // 이 부분은 위의 오류 해결시 지워주세요~
 //        return null;
 //    }
-//}
+}
